Add a check and not replace /portainer/Files/AppData/Config/prometheus/config/prometheus.yml if it already exists as a file.

+++ tools/rpi_docker_monitor.sh
@@ -21,13 +21,18 @@ echo "Creating directories..."
 sudo mkdir -p /portainer/Files/AppData/Config/prometheus/config || error "Failed to create config directory!"
 sudo mkdir -p /portainer/Files/AppData/Config/prometheus/data || error "Failed to create data directory for Prometheus!"
 sudo mkdir -p /portainer/Files/AppData/Config/grafana/data || error "Failed to create data directory for Grafana!"
-echo "Downloading Prometheus config files"
+echo "Downloading Prometheus config files if they don't exist"
 if [ -d /portainer/Files/AppData/Config/prometheus/config/prometheus.yml ];
 then
     echo "/portainer/Files/AppData/Config/prometheus/config/prometheus.yml is a directory removing"
     rm -rf '/portainer/Files/AppData/Config/prometheus/config/prometheus.yml/'
 fi
-sudo wget -O /portainer/Files/AppData/Config/prometheus/config/prometheus.yml https://raw.githubusercontent.com/oijkn/Docker-Raspberry-PI-Monitoring/main/prometheus/prometheus.yml || error "Failed to download prometheus.yml file!"
+
+if [ ! -f /portainer/Files/AppData/Config/prometheus/config/prometheus.yml ]; 
+then
+	sudo wget -O /portainer/Files/AppData/Config/prometheus/config/prometheus.yml https://raw.githubusercontent.com/oijkn/Docker-Raspberry-PI-Monitoring/main/prometheus/prometheus.yml || error "Failed to download prometheus.yml file!"
+fi
+
 if [ -d /portainer/Files/AppData/Config/grafana/grafana.ini ];
 then
     echo "/portainer/Files/AppData/Config/grafana/grafana.ini is a directory removing"
